build(lint): 头文件与浏览器口径存量清零，CI 三道 lint 收回必过语义
覆盖面先扩再清：.clang-tidy 的 HeaderFilterRegex 改写为分隔符无关（[/\]）。旧值只认
正斜杠，Windows 编译库里「正斜杠根 + 反斜杠段」的混合路径把 include/aurora/** 的头文件诊断
整片滤掉，Linux runner 全穿过——这正是「本机两口径 0」而 CI 首跑 422/470 的全部原因。扩面
一次性放出 417 条 / 64 文件，本机自此与 CI 量同一张覆盖面。

处置按「覆盖面代价」排序，未放宽任何检查规则：
- 头文件命名存量 224 → 9 → 0：命名类走 CheckOptions 按类别精确豁免（实测 clang-tidy 22 把
  类内 static inline/thread_local 数据成员归入 GlobalVariable 类，且 *IgnoredRegexp 是整串
  匹配语义：_$ 无效、.*_ 生效，只放宽大小写一维），余 9 条直接改码。
- 浏览器口径独报 39 → 35 → 0：11 处真改码 + 24 处带理由豁免。其中「仅浏览器口径命中」一族
  （dynamic-static-initializers 等）以同一份 UT、同一配置、两份编译库 A/B 复验后按口径差异
  记账，不为凑口径改配置或改码；platform.h 的 macro-to-enum 改区间式覆盖整段——命中的
  #define 随编译目标而变，逐点豁免必漂移。
- 真缺陷：WasmSurface::instances_ 由 unordered_set 改插入有序 vector，失焦回落与广播顺序自此
  确定（bugprone-nondeterministic-pointer-iteration-order 命中）；WasmSurface/WasmAriaBridge
  补显式 = delete 四件套（既成事实写进声明）。
- 生成链：generate_error_codes 后串 clang-format -style=file -i，重跑生成不再自己弄红仓库。

CI 语义：三道 lint 删去 continue-on-error 转必过，artifact 照旧（「不报」≠「跑到了」）。残余
盲区如实申报不回退：本机 native 遍在 Windows 上跑，AURORA_PLATFORM_LINUX 独享分支不进分析
（CI 默认开关下那几个 Linux-only 文件即空 TU，余下仅约百行平台分支）。

验收：native DEBUG=OFF 504 TU / DEBUG=ON 498 TU / 浏览器口径 503 TU 均 0 findings、0 断链；
native ctest 311/311、wasm ctest 309/309；format-check 836/836 幂等；check_nolint_layout
1186 条指令全合规。文档回写 CODING_STANDARDS.md §5.2、BUILD_OPTIONS.md §4.5，CHANGELOG 记
三条 fixed 与一条验收口径。

+++ include/aurora/widget/scroll_viewport.h
@@ -132,7 +132,9 @@ struct ScrollViewport {
         float candidate = 0.0F;
         switch (snap.alignment) {
             case ScrollSnapAlignment::Start:
-                candidate = std::lround(offset / ext) * ext;
+                // lround 得整条目号，转 float 属真窄化（long 32 位 > float 24 位尾数）——显式写出来，
+                // 乘法仍在 float 域做，数值与原式一字不动。
+                candidate = static_cast<float>(std::lround(offset / ext)) * ext;
                 break;
             case ScrollSnapAlignment::Center: {
                 // 条目 k 覆盖 [k·ext, (k+1)·ext)，其中心贴视口中心：offset = k·ext + ext/2 − viewport/2
